Ai-Generator/Setting-Python/requirements.txt
requests
inquirer
openai
colorama
python-dotenv
